requirements.txt
sounddevice
soundfile
numpy
hidapi
customtkinter
packaging
pillow
pynput